requirements.txt
gdown
tensorflow
streamlit
Pillow
numpy
python-dotenv
google-generativeai